Moved index_range to the module interface (that is, the include folder)
Several queries can be optimized if they receive the entire range at once. For example, querying the rank of a symbol in a wavelet tree costs 3 bitmap ranks per level. If ones wants to count the appearances of a symbol in a given index range, one has to invoke WT rank twice and subtract the returned values to get the answer. It costs a total of 6 bitmap ranks per level! However, it can be done by visiting just once each node, costing a total of 4 bitmap rank per level... A huge improvement :)

Moreover, offering a richer interface allows the user to be more expressive and allows the provider to have more optimization opportunities.